feat: reorganize skeleton components and refactor account page
- Extract skeleton components from ui/skeleton.tsx to dedicated skeletons/ folder
- Create specialized AccountOverviewSkeleton for account page loading state
- Update account page to use direct service calls with proper error handling
- Add breadcrumb navigation to account page
- Update imports across components to use new skeleton structure
- Add @/skeletons/* path alias to TypeScript configuration
- Add services type definitions to centralize service-related types
