+++ src/components/dashboard/reports/CampaignOpensLoading.tsx
@@ -7,7 +7,7 @@
  */
 
 import { Card, CardContent, CardHeader } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
+import { Skeleton } from "@/skeletons";
 
 export function CampaignOpensLoading() {
   return (
