+++ src/components/dashboard/reports/CampaignReportLoading.tsx
@@ -7,7 +7,7 @@
  */
 
 import { Card, CardContent, CardHeader } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
+import { Skeleton } from "@/skeletons";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 
 export function CampaignReportLoading() {
