+++ src/components/mailchimp/index.ts
@@ -3,5 +3,5 @@
  * Centralized exports for all Mailchimp-related components
  */
 
-export * from "./account";
-export * from "./audiences";
+export * from "@/components/mailchimp/account";
+export * from "@/components/mailchimp/audiences";
