+++ src/services/index.ts
@@ -21,6 +21,7 @@ export type { MailchimpList, MailchimpListsQuery } from "@/types/mailchimp";
 // Service factory instances
 import { MailchimpService } from "@/services/mailchimp.service";
 import { ApiServiceFactory } from "@/services/base-api.service";
+import type { ServiceName, ServiceInstance } from "@/types/services";
 
 /**
  * Get Mailchimp service instance (singleton)
@@ -39,13 +40,6 @@ export const services = {
   mailchimp: getMailchimpService,
 } as const;
 
-/**
- * Service type helpers
- */
-export type ServiceName = keyof typeof services;
-export type ServiceInstance<T extends ServiceName> = ReturnType<
-  (typeof services)[T]
->;
 
 /**
  * Helper to get any service by name
