+++ src/skeletons/components/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Generic UI Skeleton Components
+ * Reusable skeleton components for common UI patterns
+ */
+
+export * from "@/skeletons/components/Skeleton";
+export * from "@/skeletons/components/MetricCardSkeleton";
+export * from "@/skeletons/components/TableSkeleton";
+export * from "@/skeletons/components/ListSkeleton";
