+++ src/skeletons/dashboard/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Dashboard Skeleton Components
+ * Skeleton components specifically designed for dashboard layouts
+ */
+
+export * from "@/skeletons/dashboard/DashboardSkeleton";
