+++ src/skeletons/forms/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Form Skeleton Components
+ * Skeleton components specifically designed for form layouts
+ * (Currently empty - can be populated as form skeletons are needed)
+ */
+
+// Empty export to make this a proper module
+export {};
