+++ src/skeletons/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Skeleton Components
+ * Centralized exports for all skeleton loading components
+ */
+
+export * from "@/skeletons/components";
+export * from "@/skeletons/dashboard";
+export * from "@/skeletons/mailchimp";
+export * from "@/skeletons/forms";
