+++ src/skeletons/mailchimp/AccountOverviewSkeleton.tsx
@@ -1,0 +1,107 @@
+import { Card, CardContent, CardHeader } from "@/components/ui/card";
+import { Skeleton } from "@/skeletons";
+
+export function AccountOverviewSkeleton() {
+  return (
+    <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
+      {/* Account Information Skeleton */}
+      <Card>
+        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+          <Skeleton className="h-4 w-24" />
+          <Skeleton className="h-4 w-4 rounded-full" />
+        </CardHeader>
+        <CardContent>
+          <div className="space-y-2">
+            <Skeleton className="h-8 w-32" />
+            <Skeleton className="h-3 w-24" />
+            <div className="flex items-center space-x-2">
+              <Skeleton className="h-3 w-3 rounded-full" />
+              <Skeleton className="h-4 w-40" />
+            </div>
+            <div className="flex items-center space-x-2">
+              <Skeleton className="h-5 w-16 rounded-full" />
+              <Skeleton className="h-5 w-12 rounded-full" />
+            </div>
+          </div>
+        </CardContent>
+      </Card>
+
+      {/* Contact Information Skeleton */}
+      <Card>
+        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+          <Skeleton className="h-4 w-32" />
+          <Skeleton className="h-4 w-4 rounded-full" />
+        </CardHeader>
+        <CardContent>
+          <div className="space-y-1">
+            <Skeleton className="h-5 w-28" />
+            <Skeleton className="h-4 w-36" />
+            <Skeleton className="h-4 w-32" />
+            <Skeleton className="h-4 w-40" />
+            <Skeleton className="h-4 w-24" />
+          </div>
+        </CardContent>
+      </Card>
+
+      {/* Subscribers Skeleton */}
+      <Card>
+        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+          <Skeleton className="h-4 w-20" />
+          <Skeleton className="h-4 w-4 rounded-full" />
+        </CardHeader>
+        <CardContent>
+          <Skeleton className="h-8 w-20 mb-2" />
+          <Skeleton className="h-3 w-24" />
+        </CardContent>
+      </Card>
+
+      {/* Account Timeline Skeleton */}
+      <Card className="md:col-span-2 lg:col-span-3">
+        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+          <Skeleton className="h-4 w-28" />
+          <Skeleton className="h-4 w-4 rounded-full" />
+        </CardHeader>
+        <CardContent>
+          <div className="grid gap-4 md:grid-cols-3">
+            <div className="space-y-2">
+              <Skeleton className="h-4 w-20" />
+              <Skeleton className="h-6 w-32" />
+            </div>
+            <div className="space-y-2">
+              <Skeleton className="h-4 w-24" />
+              <Skeleton className="h-6 w-28" />
+            </div>
+            <div className="space-y-2">
+              <Skeleton className="h-4 w-20" />
+              <Skeleton className="h-6 w-30" />
+            </div>
+          </div>
+        </CardContent>
+      </Card>
+
+      {/* Industry Stats Skeleton */}
+      <Card className="md:col-span-2 lg:col-span-3">
+        <CardHeader>
+          <Skeleton className="h-4 w-32" />
+          <Skeleton className="h-3 w-64" />
+        </CardHeader>
+        <CardContent>
+          <div className="grid gap-4 md:grid-cols-3">
+            <div className="text-center space-y-2">
+              <Skeleton className="h-8 w-16 mx-auto" />
+              <Skeleton className="h-4 w-20 mx-auto" />
+            </div>
+            <div className="text-center space-y-2">
+              <Skeleton className="h-8 w-16 mx-auto" />
+              <Skeleton className="h-4 w-20 mx-auto" />
+            </div>
+            <div className="text-center space-y-2">
+              <Skeleton className="h-8 w-16 mx-auto" />
+              <Skeleton className="h-4 w-20 mx-auto" />
+            </div>
+          </div>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
