+++ src/skeletons/mailchimp/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Mailchimp Skeleton Components
+ * Skeleton components specifically designed for Mailchimp features
+ */
+
+export * from "@/skeletons/mailchimp/MailchimpDashboardSkeleton";
+export * from "@/skeletons/mailchimp/AccountOverviewSkeleton";
